Final definitions

+++ backend/src/models/shark_energies.js
@@ -1,9 +1,9 @@
 'use strict';
 module.exports = (sequelize, DataTypes) => {
   var shark_energies = sequelize.define('shark_energies', {
-    watth: DataTypes.INTEGER,
-    varh: DataTypes.INTEGER,
-    vash: DataTypes.INTEGER
+    watth: DataTypes.FLOAT,
+    varh: DataTypes.FLOAT,
+    vash: DataTypes.FLOAT
   }, {});
   shark_energies.associate = function(models) {
     // associations can be defined here
